// src/inventory_operation/dto/create-inventory-operation.dto.ts
import { CreateInventoryOperationDetail } from "@/inventory_operation_details/dto/create-inventory-operation-detail.dto";
import z from "zod";

// all of this is generated by API
export const CreateInventoryOperation = z.object({
  operation_type: z.enum(["LOAD", "DOWNLOAD"]),
  //document_no: z.string().min(5),
  description: z.string(),
  //total: z.number().min(0),
  //total_net: z.number().min(0),
  //total_tax: z.number().min(0),
  //total_details: z.number().min(0),
  //total_amount: z.number().min(0),
  inventory_operation_details: z.array(CreateInventoryOperationDetail).min(1),
});